% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD-SAXS.R
\name{find.xray.peak}
\alias{find.xray.peak}
\title{find peaks using the Gaussian find.peaks function from tomio.signals}
\usage{
find.xray.peak(
  data,
  sigma = 20,
  q.min = NULL,
  q.max = NULL,
  max.I = NULL,
  log.I = FALSE,
  bg.I = FALSE,
  peak.threshold = 0.05,
  peak.window = 2,
  bg.window = 80,
  verbose = FALSE
)
}
\arguments{
\item{data}{an X-ray data list}

\item{sigma}{standard deviation of the Gaussian filter
used for differentiation and smoothing
(this is in index values)}

\item{q.min}{if set, keep only q > q.min}

\item{q.max}{if set, keep only q < q.max}

\item{max.I}{cut the intensities to be below this value
useful for SAXS data, where the starting
part can span 6 orders of magnitude}

\item{log.I}{Boolean if set, take log(I) instead of I}

\item{bg.I}{Boolean if set, subtract a background from I
using the generic background function from tomio.signals
with a window parameter of bg.window}

\item{peak.threshold}{relative threshold for peaks
passed to find.peaks}

\item{peak.window}{an index value; how many points
to be used around the peak to refine the
position (passed to find.peaks)}

\item{bg.window}{the window parameter for background
subtraction, if bg.I is set.}

\item{verbose}{Boolean, if TRUE, plot the results}
}
\value{
data frame     containing the q,
                                       theta,
                                       2theta,
                                       d=2pi/w and peak intensities
}
\description{
find peaks using the Gaussian find.peaks function from tomio.signals
}
\details{
Find the peaks using the sign change of the first derivative, employing
the find.peaks() from tomio.signals package.
From the X-ray data set, use the $data$q and $data$I arrays.
}
