% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD-SAXS.R
\name{plot.xray.group}
\alias{plot.xray.group}
\title{take a variable name pattern and plot them as
X-ray data sets using plot.xray().}
\usage{
\method{plot}{xray.group}(pattern, complement = TRUE, cols = NULL, alpha = 1, legends = NULL, ...)
}
\arguments{
\item{pattern}{string, a pattern to get the variables via ls()
if an array of strings is provided, it is considered
the names of variables to load (or a hit list)}

\item{complement}{Boolean, find the complementer data set}

\item{cols}{string array, name of colors to be used}

\item{alpha}{numeric 0 - 1, a transparency value for generating colors}

\item{legends}{string array, legends to be used}

\item{...}{all further arguments are passed to plot.xray}
}
\description{
take a variable name pattern and plot them as
X-ray data sets using plot.xray().
}
\details{
Plot based on a search pattern, where the found data are all X-ray
scattering curves in the format provided by the reader files.
Especially, have a $data subset which contains q/theta and intensity
columns, suitable for plot.xray().
}
