% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD-SAXS.R
\name{plot.xray}
\alias{plot.xray}
\title{Plot a SAXS and/or WAXS (XRD) data set}
\usage{
\method{plot}{xray}(
  sp.1,
  sp.2 = NULL,
  theta = FALSE,
  drop = 3,
  add = FALSE,
  type = "l",
  xlim = NULL,
  log = NULL,
  lwd = 2,
  ...
)
}
\arguments{
\item{sp.1}{X-ray list objects}

\item{sp.2}{X-ray list objects, optional. If present the
plot is the merge of the two curves together}

\item{theta}{Boolean, whether plot 2 theta axis for x
if TRUE, log= 'y' else log='xy' by default}

\item{drop}{integer, how many points to drop from the end
of the first curve and the beginning
of the second one
ignored for single curves}

\item{add}{Boolean, if TRUE, do not make a new plot}

\item{type}{string, the type parameter of plot}

\item{xlim}{array, the xlim parameter of plot}

\item{log}{string, the log parameter of plot}

\item{lwd}{integer, the line width parameter of plot}

\item{...}{other parameters are passed to plot}
}
\value{
Nothing
}
\description{
Plot a SAXS and/or WAXS (XRD) data set
}
\details{
Plot together the spectra of two halves, ideally one is WAXS
the other SAXS.
If the second set is NULL, plot only a single curve, like
the XRD data from the XRDML reader.

In both cases take the theta or q.A.1 signals for the X-axis.
}
