% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-Xenocs.R
\name{read.xenocs}
\alias{read.xenocs}
\title{read a xenocs scattering data file}
\usage{
read.xenocs(filename)
}
\arguments{
\item{filename}{string, the file to be read}
}
\value{
a list with all parameters and data
         data is in the $data subfield as a data frame
}
\description{
read a xenocs scattering data file
}
\details{
the Xenocs system exports an angle-averaged scattering
intensity profile into a tabular text file with a large
header. This function is built to import this data for
further analysis.

## reading the parameters
the header is a set of commented lines.
These commented lines contain parameters in the form of
# key             value
the value can be text, numeric or boolean

This script tries to be non-deterministic pulling all
up, and testing conversions.

Then pull in the intensity table into a dataframe
under name: 'data'. The q values are converted to 1/nm
unit, intensity and error remains as is.
Also generate a column for theta and d equivalent values.
The column theta.rad is twice the scattering angle in radians.
(that is 'theta' is actually 2 theta)
}
