% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD-SAXS.R
\name{read.xray.files}
\alias{read.xray.files}
\title{read all data files from a folder and put them into the
main memory.}
\usage{
read.xray.files(
  folder = "./",
  filetype = c("dat", "xrdml"),
  prefix = "",
  remove = "",
  recursive = FALSE,
  envir = .GlobalEnv
)
}
\arguments{
\item{folder}{text, where to search for the files}

\item{filetype}{text, one of dat or xrdml}

\item{prefix}{text, if non-empty, prepend to variable
names with a dot}

\item{remove}{text, remove this pattern from the file name
during conversion to variable name}

\item{recursive}{Boolean, whether the searhc should be recursive}

\item{envir}{environment where to dump the results}
}
\value{
nothing
}
\description{
read all data files from a folder and put them into the
main memory.
}
\details{
Run in a similar manner as read.all.files in tomio.signals
Turn the file names to variable names such:
- remove dates as [0-9]\{4\}-\[0-9\]\{2\}-\[0-9\]\{2\}
- convert - and _ characters to dots
- add a prefix with a '.' before
}
