% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRDML-reader.R
\name{read.xrdml}
\alias{read.xrdml}
\title{read an RDXML file from Malvern Panalytical}
\usage{
read.xrdml(filename)
}
\arguments{
\item{filename}{char    the name / path to the file}
}
\value{
list with metadata and data in the $data subfield
}
\description{
read an RDXML file from Malvern Panalytical
}
\details{
read an RDXML file from Malvern Panalytical and turn it
into a list of metadata and data
To be consistent between devices, make a data dataframe
containing: I, theta, q, d for
intensity, 2Theta, q in 1/nm, and characteristic thickness
d in nm (2pi/q)
(theta is actually 2theta, double the scattering angle)
}
