% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD-SAXS.R
\name{subtract.background}
\alias{subtract.background}
\title{subtract a (measured) background signal from the signal}
\usage{
subtract.background(
  signal,
  bg,
  f = NULL,
  normalize = FALSE,
  I.N = 3,
  new.q = TRUE
)
}
\arguments{
\item{signal}{an X-ray list, where data has I for intensities}

\item{bg}{an X-ray list, where data has I for background}

\item{f}{float, scalar factor for scaling the background to the signal}

\item{normalize}{Boolean if TRUE, normalize to the zero intensities}

\item{I.N}{integer, how many points of the starting part of
intensities to use in calculating the zero
intensity? I.0 = mean(I[1:I.N])}

\item{new.q}{Boolean if we should identify the valid q-range, as where
the corrected signal is first above the background}
}
\value{
an X-ray list data$I replaced with the subtracted signal
}
\description{
subtract a (measured) background signal from the signal
}
\details{
Subtract two signals from one another: the actual signal
and one measured on the background (empty sample holder)
Normally do a 1:1 subtraction, or scale the signals to one-another
If normalize is set, then use the zero intensities to scale tha background
to the signal. (I - I.0/I.bg.0 * bg)
If 'f' as for factor is set then use I - f*bg
If normalize is set, f is ignored.
}
