% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD-SAXS.R
\name{write.xray}
\alias{write.xray}
\title{use write.table to export the intensity table of an x-ray record}
\usage{
write.xray(x, file, sep = ",")
}
\arguments{
\item{x}{the variable to be exported}

\item{file}{the file name or pointer to be exported to}

\item{sep}{separator character passed to write .table}
}
\value{
Nothing
}
\description{
use write.table to export the intensity table of an x-ray record
}
\details{
this function just dumps I, dI, q, theta, theta.rad and d into a text file,
not exporting the metadata.
It is envisioned as an easy export for further plotting or processing of
scattering curves.
}
